requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Machine Learning & Computer Vision
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.3
pillow==10.2.0

# Database
# SQLite (built-in with Python)
# For PostgreSQL: psycopg2-binary==2.9.9
# For MySQL: mysql-connector-python==8.3.0
# For MongoDB: pymongo==4.6.1

# Data Processing
pandas==2.1.4

# API Documentation
pydantic==2.5.3

# Logging and Monitoring
python-json-logger==2.0.7

# Development Tools
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0

# Optional: For advanced features
# scikit-learn==1.4.0
# opencv-python==4.9.0.80
# matplotlib==3.8.2
